Move matrix decomposition Ops into a sub-module
