+++ pytensor/tensor/_linalg/decomposition/__init__.py
