+++ tests/link/pytorch/test_slinalg.py
@@ -1,6 +1,6 @@
 import pytest
 
-import pytensor
+from pytensor.tensor._linalg.decomposition import qr
 from tests.link.pytorch.test_basic import compare_pytorch_and_py
 
 
@@ -15,6 +15,6 @@ from tests.link.pytorch.test_basic import compare_pytorch_and_py
 )
 def test_qr(mode, matrix_test):
     x, test_value = matrix_test
-    outs = pytensor.tensor.slinalg.qr(x, mode=mode)
+    outs = qr.qr(x, mode=mode)
 
     compare_pytorch_and_py([x], outs, [test_value])
