+++ tests/tensor/test_slinalg.py
@@ -15,6 +15,10 @@ from pytensor.configdefaults import config
 from pytensor.graph.basic import equal_computations
 from pytensor.link.numba import NumbaLinker
 from pytensor.tensor import TensorVariable
+from pytensor.tensor._linalg.decomposition.cholesky import Cholesky, cholesky
+from pytensor.tensor._linalg.decomposition.lu import lu, lu_factor, pivot_to_permutation
+from pytensor.tensor._linalg.decomposition.qr import qr
+from pytensor.tensor._linalg.decomposition.schur import qz, schur
 from pytensor.tensor._linalg.solve.linear_control import (
     solve_continuous_lyapunov,
     solve_discrete_are,
@@ -22,23 +26,15 @@ from pytensor.tensor._linalg.solve.linear_control import (
     solve_sylvester,
 )
 from pytensor.tensor.slinalg import (
-    Cholesky,
     CholeskySolve,
     Solve,
     SolveBase,
     SolveTriangular,
     block_diag,
     cho_solve,
-    cholesky,
     eigvalsh,
     expm,
-    lu,
-    lu_factor,
     lu_solve,
-    pivot_to_permutation,
-    qr,
-    qz,
-    schur,
     solve,
     solve_triangular,
 )
@@ -410,7 +406,7 @@ class TestSolveTriangular(utt.InferShapeTester):
     @staticmethod
     def A_func(x, lower, unit_diagonal):
         x = x @ x.T
-        x = pt.linalg.cholesky(x, lower=lower)
+        x = cholesky(x, lower=lower)
         if unit_diagonal:
             x = pt.fill_diagonal(x, 1)
         return x
